give Baby Face's Blaster +30% accuracy
This is to make the weapon viable to peel off the boss to build meter, if you want to build meter at the moment, the boss needs to be in punching range of you. This is, obviously, not really optimal.

This will very likely be increased because while this seems huge, the weapon's damage output does not change, so the speed in which a player can built up does not change, merely that the weapon has the potential to build meter when the boss is busy with someone else from close or mid-range.