Add drain (7%) over time (each second, when Scout has 20% meter) for Baby Face's Blaster
Adds drain mechanic to the Baby Face's Blaster, in which the Scout's m_flHypeMeter prop is manipulated each second. Every second 7% of the meter is drained IF the Scout has 20% or more (drains until there's 20% in the meter).

The purpose of this is to make the weapon not be able to store the speed gained from the BFB for the whole match (which is what happened in old days of Freak Fortress 2). It is very likely that the drain will be reduced OR the minimum meter will be increased, but for now these percentages are fine to not impact people who use the BFB as we want them to use it (peel, build speed, rely on ground movability rather than jumping around) while keeping the weapon as 'familiar' as possible in regular gameplay.

Adds g_flClientScoutLastHypeDrain, used to sture GetGameTime() value of when a Scout's meter was drained.

This also enables the line in which the class name of primary weapons are taken.

(note: this is not compiled, but this is basic stuff, couldnt be bothered to set up all the includes)