Prevent the Spy from receiving (mini-)crits while cloaked (non-switch)
Addresses https://github.com/Kenzzer/VSH-Rewrite/issues/7 but without turning the code into a switch-based function. Based on the source code of VSH, FF2 and VSH-2, this is pretty much guaranteed to work.

Could be turned into a switch function later.

This covers:
- The code which gives 'ever-lasting' TFCond_CritOnDamage.
- The code which gives the Spy TFCond_Buffed when wielding a revolver.
The Spy does not get anything else in the if-elseif's in vsh/client.sp, hence why one TF2_IsPlayerInCondition is added in there.